requirements.txt
Django==5.2.6
bcrypt==4.3.0
djangorestframework-simplejwt==5.5.1
djangorestframework==3.16.1
gunicorn==23.0.0
#psycopg2==2.9.10
psycopg2-binary==2.9.10
google-genai==1.36.0
deprecated==1.2.18
python-dotenv
daphne
hiredis
django-cors-headers
rest-swagger
django-filter
requests
wireup